backend/requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.28.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
langsmith>=0.1.0
# Phase 2: 地震学解析
numpy>=1.26.0
scipy>=1.12.0
websockets>=13.0
pytest-asyncio>=0.23.0
lxml>=5.0.0
datasets>=2.0.0
pytest-httpx>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
greenlet>=3.0.0
scikit-learn>=1.4.0
